Remove need for 'parameters' kwarg in iterate_jit
- For every argument to the function that is wrapped by
iterate jit, we determine if that name lives inside
the parameters for the tax calculator (or if it is the
special case of 'puf=True'). If it is, we know to treat it as if
someone had passed that name to the parameters keyword arg.

iterate_jit still accepts the 'parameters' keyword argument