Revise logic of four calculate tests to handle two changes.
The first change is the elimination of the Calculator.from_files
method; the second change is the new Calculator.__init__ method
that requires a Parameters object be passed via the params argument.
After the changes in the logic of four tests, there are no test
failures among the 78 taxcalc tests.