Use 1991 WEIGHTS in all Records constructor calls.
The 1991 PUF data are being used in the tests to emulate 2009 PUF
data, and therefore, they have a different number of filing units.  So
it is essential to use the 1991 WEIGHTS, rather than the default
weights for the 2009 PUF.  The need for doing this was highlighted by
test_Calculator_create_distribution_table and
test_Calculator_create_difference_table results.