Add "if self._current_year == 2009:" statement.
This change was make because the imputation logic in the Records class
constructor is specific to 2009, and therefore, comparing 2009 policy
parameters with non-2009 records data is inappropriate.  After adding
this one statement and indenting the imputation code, the test.py
script produces exactly the same 2009 PUF results as before adding
this one statement.  However, the test suite fails because many of the
indented lines are now longer than 79 characters, violating PEP8.
This will be fixed in a subsequent commit by moving the imputation code
into its own method.