Add Sphinx-style documentation of Record class constructor.
The new documentation contains a strong warning about the consequences
of careless use of the constructor's start_year parameter.  Also, add
more type checking of constructor parameters.  And finally, changed
SOIYR from 2008 to 2009, but this change had no effect on pytest
results or on results generated by the test.py script (using the 2009
puf.csv as Records data).  This raises the question of whether the
SOIYR variable is necessary for taxcalc logic.  It does not appear to
be used in functions.py logic.