Add missing EITC eligibility test for elderly filing units.
This test uses the imputed variable _nunextra, which is not exactly
correct for determining age-related EItC eligibility for filing units
without children.  But the approximation introduced here is likely to
be closer to the truth than simply ignoring this eligibility rule,
which is what the code was doing before this revision.  And this
approximate EITC eligibility test also has the advantage of bring the
EITC results generated by the taxcalc package in line with those
generated by Internet TAXSIM.