Add wrt_expanded_income parameter to the Calculator mtr method.
The default is of this new parameter is True, which produces exactly
same results as before the addition of the parameter.  The SimpleTaxIO
class will set wrt_expanded_income=False to facilitate cross-model
validation work with Internet TAXSIM, which does not yet compute
marginal tax rates with respect to an expanded income concept that
includes the employer-share of OASDHI payroll taxes.