Update six .taxdiffs files to reflect changes caused by PR 394.
The changes for the b1? and c1? files reflect PR 394 improvements in
the logic of the choice between standard and itemized deductions in
the face of the AMT rules.  However, the changes for the a13 and a14
files, which contain only filing units with no itemized-deduction
expenses, are puzzling.  One would think there should be no change in
the a13.taxdiffs and a14.taxdiffs files, but that is not true.  While
the total income tax liability (ovar=4) is exactly the same before and
after PR 394, the differences for four intermediate values (ovar 18,
19, 27, 28) have diverged from essentially zero to a few large nonzero
differences.  These will be investigated in the future.