Add --taxsim2441 option to simtax.py and simpletaxio.py logic.
The new option enables optional emulation of Internet TAXSIM practice
of approximating the number of Form 2441 qualified persons, for which
the the Internet TAXSIM "child care expenses" (input variable 17) are
incurred, by the total number of dependents of any age.  Not using the
option implies using the number of dependents under age 17 (Internet
TAXSIM input variable 19) in the approximation.  The actual Form 2441
rule is to qualify children only if they are under age 13, but that
is not an Internet TAXSIM input variable; hence the need to
approximate.