Update six .taxdiffs files with positive-spouse-earnings results.
Now there are 30-40 filing units with federal income tax differences
(greater than one cent) in each a1?.in file and 300-400 with income
tax differences (greater than one cent) in each of the b1?.in and
c1?.in files.  There are no differences in FICA taxes in any of the
six input files, each of which contains 100,000 filing units.