Expand verification of reform implementation in SimpleTaxIO test2.
All the new asserts are true until AMT_em for 2018 is not equal to
AMT_em for 2017, which is the year in which the indexing is turned
off.  This test failure indicates need for a bug fix.  One possibility
is that the bug is in the ParametersBase _update method which seems to
"miss" changes in indexing status of a parameter when there is no
value specified for the parameter in the year the indexing status
changes.