Fix bugs in ParametersBase _update method logic.
This commit fixes two bugs.  The current structure of the TaxBrain
webapp may be such that the following kinds of policy reforms can not
be specified, but these kinds of reforms can be specified by
constructing a reform dictionary by hand or by specifying a reform
JSON file for use by the SimpleTaxIO class.  So, it seems prudent to
revise _update so that these kinds of reforms will be implemented
correctly.

The first situation is this: a user specifies a change in the indexing
status of a parameter without specifying a value for the parameter.
This would happen when an indexed parameter becomes non-indexed.  The
old version of _update completely missed this change and left the
parameter with its original indexing status.  This commit fixes this
problem.  The expanded test2 in test_simpletaxio.py illustrates this
situation.

The second situation is this: a user specifies a change from
non-indexed to indexed and then in a subsequent year specifies an
increase in the value of that parameter.  The old version of _update
missed the earlier change in the indexing status (because it looked at
the default values, which are correct for the start year but not
necessarily for subsequent years), and so incorrectly projected the
new parameter value as a constant (instead of indexing the new value).