Add consider_imputations argument to Records class constructor.
Default value is True, which was the implied logic before the
addition, so that all code that calls the Records class constructor
will work exactly as before.  This new constructor argument has been
added to make explicit whether or not to do the variable imputations.
The addition of this argument makes the Records class constructor
argument consider_imputations analogous to the Calculator class
constructor argument sync_years.