Make Calculator ctor handle records argument same as policy argument.
The vestigial option to pass a string value for the records argument
is nowhere in the Tax-Calculator (other than in two tests) and nowhere
in TaxBrain.  As a consequence, we can remove the test_calculate
test_make_Calculator_files_to_ctor(policyfile) function and simplify
the test_make_Calculator_with_reform_after_start_year() function.