Revise FLPDYR logic along lines suggested by Dan Feenberg.
When not specifying a start_year in the Records constructor, the
current_year and FLPDYR variable are now set to PUF_YEAR, so that ages
will be correctly calculated.  Also, remove the increase of FLPDYR by
one in the increment_year() method in order to calculate ages
correctly.  These changes imply that FLPDYR no longer needs to be
included in the MUST_READ_VARS list.