Remove function arguments from functions.py that are never used.
Also, comment out local variables that are never used in the function
in which they are defined using the UNUSED-LOCAL-VAR comment prefix.

Removing unused function arguments +
  commenting out unused local variables ==> pylint score is 9.94 out of 10.

The only remaining pylint warning messages are calling attention to
TODO comments, which were placed in the code to indicate need for
additional development work.  It would be inappropriate to suppress
this pylint warning messages.