Revise IncomeTaxIO class ctor logic in light of recent Records changes.
These revisions make it easier to use inctax.py in one of two ways:
(1) with raw input data that are adjusted in any way (that is, no
imputation and no blowup or aging), or (2) with proprietary 2009
puf.csv data that is imputed and blown up in the standard manner.  The
first way may be useful in cross-model-validation work, while the
second way may be useful in debugging Tax-Calculator income tax logic.