Remove cmbtp-calc and earnings-split logic from Records class.
This removed logic is now in the taxdata repository's
final-preparation script, so it is being executed once, when the
puf.csv file is created, instead every time the Tax-Calculator
is executed.

All the unit tests pass with this commit (using the new puf.csv file)
except for minor changes in the marginal tax rates histograms and
the failure of the test_percentage_change_gdp in the
test_macro_elasticity.py file.  This later failure is probably
caused by the use of old 1991 data that has not been prepared
in the new manner.